vm-connector: run salt shim in environment with dummy scp
Salt shim checks if scp binary is present, because normally it's used
for deploying salt minion. In Qubes we simulate scp with qvm-copy-to-vm,
so it isn't needed. To not require additional package installed (for
example in minimal template), launch salt shim in environment with dummy
scp present.

Fixes QubesOS/qubes-issues#2059