ssh-wrapper: don't use qvm-copy-to-vm
Since the intention of 'scp' here is to:
 - override target file (if any)
 - place it in specific directory, outside of QubesIncoming
it doesn't make sense to use qvm-copy-to-vm, as it adds more troubles
than good. Simple 'cat' is enough.
This solves problem when QubesIncoming is not cleaned up after failed
salt invocation, preventing further attempts (becuse qvm-copy-to-vm
refuse to override existing files).

QubesOS/qubes-issues#2600