Revert "Force 'systemd' service module in dom0"
This reverts commit 80ea2ea62a3ab98b061fb77b53b8d7a8752c755d.

Salt 2015.5 crashes if 'providers' setting is specified. This is fixed
in newer salt version, but we won't have that in fc23 dom0.
So, drop this setting (making 'service' state module broken again).
Upstream report of this problem but in different environment:
saltstack/salt#24350

This wasn't an issue earlier, as minion.dom0.conf was ignored.