Check for targeted states in parallel
This moves the check for a qube having targeted states into the
parallelly run `run_one` function. Usually, if a local state
is applied (during a highstate on the AdminVM), the call to
`state.show_top` would return an error, but this can be overridden
with `concurrent=True`.

Also uses the current interface to the caller (`cmd`, not
`function`) since the latter causes invalid kwargs to be passed.

Loading opts repeatedly vs copying and passing them did not make
a difference in testing.