Ensure custom modules can run on salt-ssh master
Usually, salt-ssh is run from a Salt master with a permanent cache dir.
In Qubes OS, a new DispVM is created for each invocation. Custom modules
need to be synced from the fileserver to the cache dir to be available
to Salt. For salt-ssh targets, this is done automatically for relevant
module types.
Modules that need to run on the "master" on the DVM on the other hand
(tops, outputters, wrappers) require a synchronization before they are
available.