Allow syncing all module types via qubesctl
The master on the mgmt DVM needs to have a synced extmod cache before
invoking salt-ssh, otherwise custom modules that need to run there will
not be available.
Only the saltutil runner will sync master-side modules, which is
untrivial to call on (by default minion-only) dom0 and would sync to a
separate cache dir by default. This provides a built-in way to include
master-only module types into the minion extmod cache, which will then
be transferred to the mgmt DVM as the synced master cache.