Support 'state.apply' in addition to 'state.highstate'
The former is more logical name and is preferred upstream. Handle it in
the qubesctl command, to apply states only to VMs where there is
anything to apply.

It's a bit more complicated, becase (contrary to state.highstate) it can
be also used to apply a single state not using top file. So, apply
optimization only when 'state.apply' is the last word. This will
correctly handle simple cases like `qubesctl --all state.apply`, but if
one use some extra arguments like `qubesctl --all state.apply test=True`
it will disable the optimization and will call it on all the VMs (also
where there is nothing to apply). Since full command parsing will be
complicated (and keeping it in sync with salt will be fragile), opt on
the safe side - better have possibly suboptimal performance but correct
result, instead of risking incorrect results.

Related to QubesOS/qubes-doc#1266