Fix handling empty base top
When the base top is empty, the "if tops" condition fails even in the
DispVM where tops.yaml file exists. This leads to trying salt.top in
a VM, which isn't installed there.

Fix this by using `is_text_file` test instead of checking for emptiness.
But then, merge_tops() can't handle empty string - it expects a proper
list of top files. To fix this, set tops to empty list if it's an empty
string.

Fixes QubesOS/qubes-issues#8491